Updating provision and destroy for aro to handle partial provisions

+++ aro/destroy.sh
@@ -7,9 +7,6 @@ RED='\e[31m'
 YELLOW='\e[33m'
 CLEAR='\e[39m'
 
-# Ensure we fail out if something goes wrong
-set -e
-
 #----LOAD ENV VARS FROM INPUT FILE----#
 if [ -z "$1" ]; then
     printf "Usage: ./destroy.sh <path-to-json-file>\n"
@@ -29,6 +26,9 @@ AZURE_BASE_DOMAIN=$(cat $1 | jq -r '.AZURE_BASE_DOMAIN')
 USERNAME=$(cat $1 | jq -r '.USERNAME')
 PASSWORD=$(cat $1 | jq -r '.PASSWORD')
 CONSOLE_URL=$(cat $1 | jq -r '.CONSOLE_URL')
+
+# Set a skipped flag that we'll set if we skip any deletions.  Used to skip deletion of state files.
+skipped=0
 
 
 #----VALIDATE ENV VARS----#
@@ -73,7 +73,7 @@ if [ "$?" -ne 0 ]; then
     exit 1
 fi
 
-if [ ! -z "$AZURE_SUBSCRIPTION_ID" ]; then
+if [[ "$AZURE_SUBSCRIPTION_ID" && "$AZURE_SUBSCRIPTION_ID" != "null" ]]; then
     az account set --subscription $AZURE_SUBSCRIPTION_ID
     if [ "$?" -ne 0 ]; then
         printf "${RED}Unable to set azure subscription, az account set --subscription $AZURE_SUBSCRIPTION_ID returned non-zero.${CLEAR}\n"
@@ -94,23 +94,41 @@ printf "${CLEAR}"
 
 
 #----DELETE NS, ARO CLUSTER, VNETS, AND DNS ZONES----#
-printf "${BLUE}Deleting the resource group ${RESOURCE_GROUP_NAME} and related DNS Zones, VNets, and ARO Cluster ${CLUSTER_NAME}.${CLEAR}\n"
-printf "${YELLOW}"
-az group delete -n ${RESOURCE_GROUP_NAME} -y;
-printf "${CLEAR}"
+if [[ "${RESOURCE_GROUP_NAME}" != "null" ]]; then
+    printf "${BLUE}Deleting the resource group ${RESOURCE_GROUP_NAME} and related DNS Zones, VNets, and ARO Cluster ${CLUSTER_NAME}.${CLEAR}\n"
+    printf "${YELLOW}"
+    az group delete -n ${RESOURCE_GROUP_NAME} -y;
+    printf "${CLEAR}"
+else
+    printf "${RED}Couldn't find RESOURCE_GROUP_NAME key in ${1}, skipping resource group deletion.${CLEAR}\n"
+    skipped=1
+fi
 
 
 #----DELETE BASE DOMAIN RECORD SETS---#
-printf "${BLUE}Deleting the record sets for the cluster base domain in the ${AZURE_BASE_DOMAIN_RESOURCE_GROUP_NAME} resource group.${CLEAR}\n"
-printf "${YELLOW}"
-az network dns record-set ns delete \
-    -g ${AZURE_BASE_DOMAIN_RESOURCE_GROUP_NAME} \
-    -z ${AZURE_BASE_DOMAIN} \
-    -n ${CLUSTER_NAME} \
-    -y;
-printf "${CLEAR}"
+if [[ "${AZURE_BASE_DOMAIN_RESOURCE_GROUP_NAME}" != "null" && "${AZURE_BASE_DOMAIN}" != "null" ]]; then
+    printf "${BLUE}Deleting the record sets for the cluster base domain in the ${AZURE_BASE_DOMAIN_RESOURCE_GROUP_NAME} resource group.${CLEAR}\n"
+    printf "${YELLOW}"
+    az network dns record-set ns delete \
+        -g ${AZURE_BASE_DOMAIN_RESOURCE_GROUP_NAME} \
+        -z ${AZURE_BASE_DOMAIN} \
+        -n ${CLUSTER_NAME} \
+        -y;
+    printf "${CLEAR}"
+else
+    printf "${RED}Couldn't find AZURE_BASE_DOMAIN_RESOURCE_GROUP_NAME and AZURE_BASE_DOMAIN in ${1}, skipping record set deletion.${CLEAR}\n"
+    skipped=1
+fi
 
 
 #----SUCCEED IF WE MADE IT THIS FAR----#
-printf "${GREEN}Successfully deleted the ARO Cluster ${CLUSTER_NAME} and related resources.  Deleting metadata file.${CLEAR}\n"
-rm -f $1
+if [[ "$skipped" ]]; then
+    printf "${GREEN}Successfully deleted the ARO Cluster ${CLUSTER_NAME} and related resources.  Deleting state and creds files.${CLEAR}\n"
+    rm -f $1
+    if [[ -f "${CLUSTER_NAME}" ]]; then
+        rm -f ${CLUSTER_NAME}.creds.json;
+    fi
+else
+    printf "${RED}Some details were missing from the state file and were skipped, we won't clean up $1 or ${CLUSTER_NAME}.creds.json.${CLEAR}\n"
+    printf "${BLUE}If you were trying to clean up a partial failed provision, this probably isn't a problem, so you can delete $1 and ${CLUSTER_NAME}.${CLEAR}\n"
+fi
